Fixed some bugs, cleaned up algorithm code.
