+++ next/apps/Apps/CardinalBanditsPureExploration/CardinalBanditsPureExploration.yaml
@@ -41,7 +41,7 @@ initExp:
               alg_id:
                 description: A supported algorithm type in the system.
                 type: str
-                values: [RoundRobin,LilUCB,LilUCB_fast]
+                values: [RoundRobin,LilUCB]
               alg_label:
                 description: A name for this algorithm of your choosing.
                 type: str
