+++ next/apps/Apps/CardinalBanditsPureExploration/algs/LilUCB/LilUCB.py
@@ -12,15 +12,18 @@ class LilUCB(CardinalBanditsPureExplorationPrototype):
 
   def initExp(self,butler,n,R,failure_probability,params):
     butler.algorithms.set(key='n', value=n)
-    butler.algorithms.set(key='failure_probability',value=failure_probability)
+    butler.algorithms.set(key='delta',value=failure_probability)
     butler.algorithms.set(key='R',value=R)
-    arm_key_value_dict = {}
-    for i in range(n):
-      arm_key_value_dict['Xsum_'+str(i)] = 0.
-      arm_key_value_dict['X2sum_'+str(i)] = 0.
-      arm_key_value_dict['T_'+str(i)] = 0.
-    arm_key_value_dict.update({'total_pulls':0})
-    butler.algorithms.increment_many(key_value_dict=arm_key_value_dict)
+
+    empty_list = numpy.zeros(n).tolist()
+    butler.algorithms.set(key='Xsum',value=empty_list)
+    butler.algorithms.set(key='X2sum',value=empty_list)
+    butler.algorithms.set(key='T',value=empty_list)
+
+    priority_list = numpy.random.permutation(n).tolist()
+    butler.algorithms.set(key='priority_list',value=priority_list)
+
+    butler.algorithms.increment(key='total_pulls',value=0)
 
     return True
 
@@ -28,62 +31,86 @@ class LilUCB(CardinalBanditsPureExplorationPrototype):
   def getQuery(self,butler,participant_dict,**kwargs):
     do_not_ask_hash = {key: True for key in participant_dict.get('do_not_ask_list',[])}
 
-    key_value_dict = butler.algorithms.get()
-    n = key_value_dict['n']
-    sumX = [key_value_dict['Xsum_'+str(i)] for i in range(n)]
-    T = [key_value_dict['T_'+str(i)] for i in range(n)]
-
-    R = key_value_dict['R']
-    delta = key_value_dict['failure_probability']
-
-    T = numpy.array(T)
-    mu = numpy.zeros(n)
-    UCB = numpy.zeros(n)
-    A = []
-    for i in range(n):
-      if T[i]==0:
-        mu[i] = float('inf')
-        UCB[i] = float('inf')
-        A.append(i)
-      else:
-        mu[i] = sumX[i] / T[i]
-        UCB[i] = mu[i] + numpy.sqrt( 2.0*R*R*numpy.log( 4*T[i]*T[i]/delta ) / T[i] )
-
-    if len(A)>0:
-      priority_list = numpy.random.permutation(A)
-    else:
-      priority_list = numpy.argsort(-UCB)
+    priority_list = butler.algorithms.get(key='priority_list')
     
+    A = []
     k = 0
-    while k<len(priority_list) and do_not_ask_hash.get(priority_list[k],False): 
+    while k<len(priority_list):
+      if not do_not_ask_hash.get(priority_list[k],False):  
+        A.append(k)
+        if len(A)>=10: # tradeoff number: the larger it is the less likely it is to have collisions, but the algorithm is not as greedy. The delay between updates in processAnswer should be considered when setting this number
+          break
       k+=1
-    if k==len(priority_list):
+    if len(A)==0:
       index = numpy.random.randint(n)
     else:
+      k = numpy.random.choice(A)
       index = priority_list[k]
 
     return index
 
   def processAnswer(self,butler,target_id,target_reward): 
-    butler.algorithms.increment_many(key_value_dict={'Xsum_'+str(target_id):target_reward,'X2sum_'+str(target_id):target_reward*target_reward,'T_'+str(target_id):1,'total_pulls':1})
-    
+    butler.algorithms.append(key='S',value=(target_id,target_reward))
+
+    if numpy.random.rand()<.1: # occurs about 1/10 of trials
+      butler.job('update_priority_list', {},time_limit=5)
+
     return True
 
   def getModel(self,butler):
     key_value_dict = butler.algorithms.get()
+    R = key_value_dict['R']
     n = key_value_dict['n']
-    sumX = [key_value_dict['Xsum_'+str(i)] for i in range(n)]
-    T = [key_value_dict['T_'+str(i)] for i in range(n)]
+    sumX = key_value_dict['Xsum']
+    sumX2 = key_value_dict['X2sum']
+    T = key_value_dict['T']
 
     mu = numpy.zeros(n)
+    prec = numpy.zeros(n)
     for i in range(n):
       if T[i]==0 or mu[i]==float('inf'):
         mu[i] = -1
+        prec[i] = -1
+      elif T[i]==1:
+        mu[i] = float(sumX[i]) / T[i]
+        prec[i] = R
       else:
-        mu[i] = sumX[i] / T[i]
-
-    prec = [numpy.sqrt(1.0/max(1,t)) for t in T]
+        mu[i] = float(sumX[i]) / T[i]
+        prec[i] = numpy.sqrt( float( max(1.,sumX2[i] - T[i]*mu[i]*mu[i]) ) / ( T[i] - 1. ) / T[i] )
     
-    return mu.tolist(),prec
+    return mu.tolist(),prec.tolist()
+
+  def update_priority_list(self,butler,args):
+    S = butler.algorithms.get_and_delete(key='S')
+
+    if S!=None:
+      doc = butler.algorithms.get()
+
+      R = doc['R']
+      delta = doc['delta']
+      n = doc['n']
+      Xsum = doc['Xsum']
+      X2sum = doc['X2sum']
+      T = doc['T']
+
+      for q in S:
+        Xsum[q[0]] += q[1]
+        X2sum[q[0]] += q[1]*q[1]
+        T[q[0]] += 1
+
+      mu = numpy.zeros(n)
+      UCB = numpy.zeros(n)
+      for i in range(n):
+        if T[i]==0:
+          mu[i] = float('inf')
+          UCB[i] = float('inf')
+        else:
+          mu[i] = Xsum[i] / T[i]
+          UCB[i] = mu[i] + numpy.sqrt( 2.0*R*R*numpy.log( 4*T[i]*T[i]/delta ) / T[i] )
+
+      priority_list = numpy.argsort(-UCB).tolist()
+
+      butler.algorithms.set_many(key_value_dict={'priority_list':priority_list,'Xsum':Xsum,'X2sum':X2sum,'T':T})
+
