--- next/apps/Apps/CardinalBanditsPureExploration/algs/LilUCB_fast/__init__.py
@@ -1,1 +1,0 @@
-from .LilUCB_fast import *
